pyproject.toml
[project]
name = "chart-sayer"
version = "0.1.0"
description = ""
authors = [
    { name = "openhands", email = "openhands@all-hands.dev" },
    { name = "lowhung", email = "hounslow@algalonlabs.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "discord-py (>=2.5.2,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openai (>=1.70.0,<2.0.0)",
    "python-telegram-bot (>=22.0,<23.0)"
]


[build-system]
packages = [{ include = "src" }]

requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
